requirements.txt
pathway[all]>=0.11.0
python-dotenv==1.0.1
mpmath==1.3.0
kafka-python==2.0.2
praw==7.7.1
requests==2.31.0
Pillow==10.3.0
Requests==2.31.0
openai==1.50.2